feat: Refactor profile DTOs and update imports
- Renamed profile.dto.ts to profile.ts for consistency in naming.
- Updated import paths in profile.controller.ts, profile.service.ts, and user.repository.ts to reflect the new file structure.
- Introduced a comprehensive Profile DTO with various nested classes for better data handling and validation.
